Numerical derivatives
Implementation is rather general and allows one to use them not only for debugging but also for real calculation. A keyword NUMERICAL_DERIVATIVES can be used in directives such as COLVARs or FUNCTIONs, and automatically performs numerical derivatives instead of analytical ones. Moreover, I added a DUMPDERIVATIVES directive which prints on a file the derivatives of the (ARG) colvars. E.g.:
DISTANCE LABEL=C1 ATOMS=0,1 COMPONENTS
DISTANCE LABEL=C1n ATOMS=0,1 COMPONENTS NUMERICAL_DERIVATIVES
DUMPDERIVATIVES ARG=C1,C1n FILE=derivC1
will define two CVs, with the same value, but one with analytical and the other with numerical derivatives, and will write two columns on file "derivC1" with the derivatives, which can be compared for debugging.

BE CAREFUL: I did not test it with non-orthorombic cells, there may be problems in the virial components