Added possibility to enforce numerical derivatives
There is a new method for colvars and function, enforceNumericalDerivatives(),
which can be used in the constructor so as to enforce numerical derivatives.
It should be added to complex colvars for which we do not want to implement
numerical derivatives, or can be used at preliminary stage during CV
implementation.

I also added a check for colvar ENERGY, which cannot be done with
numerical derivatives