Change in the workflow
Now all operations to retrieve atoms and to apply forces
are done by the ActionAtomistic class. This implies transfering
a bit of the complexity in Atoms to ActionAtomistic, which
is good. Also, it forces to follow a simpler scheme,
where each ActionAtomistic is requiring a single list
of atoms.

The advantage is that atoms are retrieved in usual Action
order (input file), so that it is easier to follow what the
code is doing. This also opens the way to Actions that
can modify the stored atomic positions for subsequent
Actions.