Tentative class WholeMolecules
This is the equivalent of the old ALIGN_ATOMS directive.
It is more flexible in that it admits multiple lists of linked
atoms and can be applied multiple times.

I also added a GenericDumpAtoms class to write down a .xyz of
the stored (modified) positions, just to debug WholeMolecules.
We may extend it to other formats (e.g. .pdb or .gro).

(see manual)