Added methods to export/import all atoms to/from a string
It is a very primitive "serialization" interface. I just put it here
so as to use it with replica exchange to send atoms to other replicas.
Still, we should plan something more general than this and replace it,
especially when we will need to pass between replica more complex objects.