Splitting performCalc into pieces
In particular, I would like to be able to just do the
"forward" loop (calculate()) and not the "backward" loop
(apply()). In this manner, when computing foreign biases
for acceptance in replica exchange, I will only call the
forward loop, which computes colvars and biases. Clearly,
we should move printing/analysis/hills-adding stuff
outside the "calculate" method, so as to avoid doing these
things in the acceptance calculation. I wait for this further
cleaning since, until replica exchange is not actually there,
there is no practical difference. Also, we may decide to change
a bit the name of methods (calculate/apply are becoming a bit
misleading?)