Added tool to add a suffix to all open files
It is now possible to automatically add a suffix to
all the files opened through the Action::fopen tool.
Opening is done as follows:
First the file+suffix is tried.
In case of error the file without suffix is tried.

In this manner, it should be possible to provide a single
input file also for replica exchange stuff. For output files,
it should make any difference.

I also added a stupid keyword to test this feature. Just use
_SET_SUFFIX pippo
in the plumed.dat file to add a suffix pippo to opened files