Added tool to recover total bias
This is useful to compute acceptance.
I did it in the "dirty" way: if a value named "bias" exist inside any
action, it is accumulated into the total bias. I also changed the name
of values in Restraint and MovingRestraint so that the energy is now called
"bias" and non-capital letters are more consistent