Fixes
Removed the loader for runtime linking from PlumedMain.cpp, so as
to ease reading of that file. Now, registering is done from
PlumedMainInitializer. Also, I removed the wrapper to dlopen() and used
dlopen directly. This allows to specify better its options. In particular,
I used RTLD_LOCAL for loading of collective variables which should
allow a clearer separation of loaded modules.