Added stuff for taking care of Boltzmann constant
This is the constant to transform temperature into energy,
so it is only needed when dealing with temperatures.
Since energy is already expressed in arbitrary units, I
did not find an elegant way to fit also this into the Unit
object. I left only two choices: physical value,
which is hardcoded and it's the default, and natural units,
which means that temperature is measured in energy units (kb=1).
To set the latter, from MD use cmd("setNaturalUnits"),
from plumed use "UNITS NATURAL" in the input file.

Notice that in this manner, even for a LJ code
one still needs to say "UNITS NATURAL" in the input. This
is to keep consistency with the fact that units in the input are
declared in the input and units for MD are declared in MD.