Added collective variables: angle and torsion
I tried to keep the tools to compute them separate from the
colvar implementation, so that it will be possible to use them
for other colvars.  For torsion still I am using numerical derivatives!
This clearly has to be fixed.

I debugged them against plumed1. Also, for angle, I checked the
derivatives comparing with the numerical ones and I still
have the usual problem with some off-diagonal contribution to
cell derivatives when applying pbc. There might be some hidden
bug there to be fixed...