Added friend versions of determinant, inverse and transpose
E.g., determinant can be computed as determinat(A) instead of
A.determinant(). It seems to me that syntax is more clear in this
manner.

I also added an extProduct operation which given two vectors
gives their external product. This is equivalent to Tensor(v1,v2),
but I think its syntax is clearer.