Removed "using namespace std" statements in .h files
It is dangerous to use "using namespace std" in .h files, since this implies that all the sources
including these .h files will explicitly use the std namespace. This could create problems of name
collision (e.g. if we use names which were already used in the std library).

On the other hand, in .cpp files we can do whatever we want, as they are not included in other files.