New makefile
I was tired of waiting for all dependencies to be generated
evey time. Now dependencies are generated during compilation,
and only for those file that changed.
The only thing that I do not like is that for AIX there is
a special case: indeed the option for the compiler that
generates dependencies also compiles in that case! On the other
hand with gcc/intel there are two steps: one to generate dependencies
and another to compile