Added framework to add command line tools
I tried to do it in such a manner that everything fits
into the plumed library (but for a trivial "main"),
so that all the command line tools can indeed be called as if they
were routines.

I used a base class CLTool which can be inherited from, similarly
to the Action class. I also temporarily copied ActionRegister
stuff into a CLToolRegister which allows registering of tools
at startup. This allows for a very flexible extensibility: it
is sufficient to load runtime a library containing a CLTool and
it becomes available.