Added possibily to run the CLTools in parallel
I added the communicator as an argument of main()
function in CLTool.h. This means that each tool
will be runnable with MPI. Notice that
the communicator defaults to MPI_COMM_WORLD,
which is the reasonable choice. Still I think it is
useful to allow passing the communicator if one wants
to run with MPI several tools in parallel.