Added --no-mpi option
plumed --no-mpi does not initialize MPI.

It could be useful to avoid MPI initialization when
running serially. However, on sp6@cineca, it seems that
startup spees is not sensitive to this parameter: the only
way to have an executable which immediately runs is to
compile it without MPI! Still I leave this option here.