Command line tools now use the standard Plumed.h interface
I modified the command line tools stuff so that it is now possible
to access it using the Plumed.h interface (i.e. cmd(...)). This
allows to call the command line tools directly from other languages,
and is more consistent with the way plumed library is accessed from
MD codes. Notice that now also a plumed-runtime executable
is produced, which looks for PLUMED_KERNEL variable at runtime
to choose the proper library to load. Since all the command line
tools are built inside the library, this just means that a
different plumed version is used (similar to runtime mode
for patching MD codes).

Moreover, I hard-coded the PLUMED_ROOT variable in the
PlumedConfig.h header file. In this manner, it is sufficient to
have plumed in the path.