Added check for wrong words on the input line
I added a call to "checkRead" which is done after complete creation
of an Action object. In this manner, it is not necessary to do it in
the Action constructor.

I did it as I realized that checkRead() was missing in the GenericGroup
constructor, and I thought that it is better to enforce it here.