Implemented parenthesis in parser
It is now possible to use parenthesis (by default "round" ones)
in the input. The way parenthesis are interpreted is the following:
Whenever an open parenthesis is found, separators (i.e. spaces) are not
taken into account until the corresponding closing parenthesis is found.
This allows to embed spaces in a single "word".

Moreover, I added "space" as a separator for parseVector. In this manner,
it is possible to specify e.g. groups of atoms alternatively as
ATOMS=1,2,3
ATOMS=(1,2,3)
ATOMS=(1 2 3)

Finally, I  modified a regtest to include parenthesis