Extended switching functions
More general switching functions are available now for coordination.
Documentation is coming (I would like to discuss with Gareth possible
automatic documentation first).

Presently the selection among switching functions is done with an
enum. An abstract base class would be better, but would require
to change the way switching functions are used (they should be
used by pointer and not by value). Since they are alreay used
in many places, I prefer to wait for this.