Flexible hills
This is the adaptive hills scheme that is called in the metadynamics just by using
ADAPTIVE=GEOM SIGMA=0.1   (only one sigma, in this case has units of distance)
or
ADAPTIVE=DIFF SIGMA=10.   (only one sigma, in this case has units of decay time)

the output in HILLS changes accordingly and presents MV (multivariate) that stands
for a band-representation of the cholesky-decomposed sigma^2
(in which at the beginning you have the diagonal terms that should give you an idea
of how wide are the hills).