License related stuff
I added a PEOPLE file containing list of authors and a script
(src/header.sh) which applies the Licence to all the source files.

It can be applies multiple times (it automatically deletes the present header
and replace it with the new one). Thus, to change the header, edit header.sh
and execute it.