Fixed two bugs affecting numerical derivatives and virial
One bug was in Pbc: tolerance for considering boxes as orthogonal
was too loose and was not correctly interpreting orthorhombic
boxes shifted by numerical derivatives.

The other was in numerical derivatives, and was wrong transpose
matrix. It was only affecting cell components of
numerical derivatives with nonorthorombic cells

I also fixed all the regtests