PlumedFile tool implemented
Documentation is still a bit poor. I used it to replace
the Log implementation and also for in the GenericPrint class.
In the log, this implies a small change (FIELDS line is
written in also in the log).

I tried to make everything so that regtests are not affected. Only,
change, now the rt17 (using _ROTATE) works correctly, writing
field names as they are updated.