Implementation of RESTART
Using a RESTART keyword in the input, files are automatically
appended. When not used, files are backed up (a la gromacs).

Notice that this works only when using open("file"). The rationale
is that if you say open("file","a") or open("file","w") then you
know what you are doing. I also removed the extra mode specifier
for PlumedIFile (it was redundant: always "r").

If the behavior is not clear enough we can change it...