Fixed license for simplemd (fortran version)
I added a statement about the fact that the fortran version of simplemd
is not part of plumed. I prefer to do this as I already gave it around
to many students and it is on my website since 4 years or so. In the end,
we do not even need this code as we have the C version embedded in plumed.