Stuff for templates and changes to developer + user manuals
Added a tool for generating template versions of the various collective
variables.  This involved changing the secondary structure variables
work (by default you now do LESS_THAN).  In addition, plumed now understands
that atoms can be specified in multiple ways sometimes.

Also changed the developer manual so as to make it easier to find the
tutorials