Changed the way the cv in a part of the box is calculated.
I have tried to make this multiColvar more flexible.  In essence
the average value of the CV is calculated in a region of fixed volume.
This volume can be specified in terms of the cell box and I am planning
to add other ways of specifying the volume presently.  The old version
of this particular multicolvar thing was actually not working so as well
as being a rework this is a fix