Added periodicity data to output in colvar and hills files
Periodicities of values are stored as strings inside Values.
These strings are output as constant fields inside the colvar and
hills files.  In addition the fact that we now use strings means that
when you have a function and you have to specify its periodicity you can
use PERIOD=-1.5pi,1.5pi or some other thing like that