Optimized PBCs
I took advantage of the fact that we have the reduced cell. This
allows to avoid the comparison of different shifts for
many values of the scaled coordinates, as well as using the
sign of the scaled coordinates to check which shifts should
be attempted. The average number of shift vector to be evaluated
is approximately 1 (plus the "null-shift")

There is also a Pbc::test() function which I used during debugging.
I leave it there for future reference.