SQUASHING COMMITS for sum_hills except regtest
This implements sum_hills tool via command line
it is called via a cltool that creates a
fake set of CVs and then creates an function actionwitharguments
which acts as hills integrator in the initial phase
(no calculate phase. It dies immediately). Hypothetically
this structure should allow also an integration  of the hills
while doing the simulations (not yet implemented).
It makes use of the Kernels instead of the internals of metad.
(TODO: replace meta with kernels)
See the documentation on cltool/SumHills.cpp

So it implementes:
added FakeCv to be used with sum_hills
sum hills accepts  stride (--stride) projections (--idw) via names
automatically grid sizing (beware, sometimes is crazy),
histogram correction for multivariate gaussians (--histo)
A one-line gaussian reader for metadynamics (more flexible)
A BiasRepresentation class which is intended to keep the list of kernels and
the grid representation in a single bunch. Can be useful when replacing
metad with kernels.
TODO: when periodic automatically sets the boundary. this must be fixed in a more
clever way
TODO  when projecting handle the derivatives so to have a grid that is usable for restaring on a
lower dimension