Install procedure
It is now possible to install plumed with "make install".

Installation is done just copying most files (e.g. obj needed
for patching) and fixing hardcoded paths. The advantage is
that it is possible to fix a location with environment variables
PLUMED_PREFIX (e.g. /opt/local/) and a suffix with
PLUMED_LIBPREFIX (e.g. 2.0.0). This will help in installing multiple
plumed versions in the same path.

Moreover, also include files are copied to e.g.
/opt/local/include/plumed, so that it is possible to use the plumed
tools library as a normal system library