Changed it so that you can use less_than/more_than more than one time per line
In addition I made some small changes to Vessel base class.  When numbered keywords
(WITHIN1, WITHIN2... )  are read in the number (the 1, 2... ) is also passed to the
vessel.  This makes it easier to create different value names for the various values
calculated by different instances of vessels of the same type.