Added multicolvar way of doing DHENERGY.
The multicolvar way of doing DHENERGY is particularly useful if you want
to calculate multiple instances of the DHENERGY with different parameters.
By using the multicolvar you can avoid the recalculation of the distances
for each different parameter set. You can also use the multicolvar to calculate
COORDINATION (LESS_THAN) and the DHENERGY for two groups of atoms at the same time