Moved some of the neighbor list stuff from ActionWithVessel to MultiColvar
This is an attempt to make it easier to re-use ActionWithVessel for things other than
MultiColvar. There are multiple ways to do a neighbor list and so I felt this functionality
should be separated somewhat.  In addition the way I had implemented the neighbor list for
multicolvar meant that the list was being updated one step later than it should have been.
Also did some tidying of the multicolvar and vessel code