A big reworking of multicolvars that makes the naming of the output values easier to understand
I have greatly simplified the hierarchy of classes that inherit from Vessel.
Furthermore, I have re-used much of the parsing functionality in Action so as
to make adding a Vessel more like adding an Action. All the changes in regtests are
just because of changes of the names in the header line.

In addition I added functionality so that multicolvar can pretend to be colvars
Multicolvars have value that are always labelled as label.something.
However, in AlphaRMSD etc the default is to calculate the number of
RMSDs that is less than a certain target value.  Furthermore, in the
manual these MultiColvars are listed as colvars.  The multicolvar functionality
inside alpharmsd being a feature for "advanced users".  Hence, to avoid confusion
I added something so that by default the value in alpharmsd etc is just
label.