Changed manner to choose RMSD version
TYPE=OPTIMAL is the old version
TYPE=OPTIMAL-FAST is the new version

Notice that the new version only works with align==displace.
Also notice that old/new version give inconsistent results when weights which are not all equal (or zero) are used. This is something that will be fixed
soon in the old version.

The new version is only used in a couple of regtests, since it still have some problem of numerical stability