Made it so manual wont compile if plumed is not compiled
Two users have complained to me that the manual is rubbish.  I subsequently
found out that they had built the manual without first building the code. So
they didn't have any of the keyword descriptions in the manual that compiled
This manul now can't be compiled if the code isn't compiled.  It seems that
having lots of error messages generated when the manual compiles wrongly is
not sufficient for some people!