Implemented fast weighted alignment with align!=displace
In this version the fast routine is hard-coded. I do not compute explicitly the derivatives of the rotational matrix.
Instead, I assume them to be zero and at the end add a contribution which resets total "force" and "torque" (i.e.
I enforce translational and rotational invariance of the RMSD). In test rt47/check one can appreciate that
numerical derivatives come out correct. I do not know exactly why...