Improvements to the way that multicolvars in a subregion of the box are calculated
Previously it was only possible to calculate the average value of the multicolvar
in a part of the box.  Now you can calculate the distribution inside the box.
This means you can use the functionality in ActionWithVessel to calculate the minimum,
the number of colvars less than, number of colvars more than and so on.

Small changes to core headers are changes to make certain routines virtual so that I
can replace them in derived classes