Added dynamic lists for central atom derivatives
This should make the code a bit faster as we don't have to loop over all the atoms
any more when using the position of a central atom.  We only have to loop over the
positions of atoms that have a central derivative (for most things 1 atom).  Also
changed DynamicList so we aren't doing lots of vector.push_backs