Changes to make multicolvar faster when dealing with coordination number like variables
The code now only adds derivatives for the quantities being accumulated
when the values are greater than TOL.  In addition, all derivative
accumulation can now be done in vesselbase. (MergeDerivatives is still in
MultiColvar but only because this makes it faster.)  This makes it easier to add new
object that inherit from ActionWithVessels.  Lastly I split the two
parts of apply in MultiColvar to remove code duplication.