Fixes to neighbor lists in ActionWithVessel
ActionWithVessel is calculating sums.  For many of the quantities
ActionWithVessl calculates (e.g. less_than, more_than, coordination numbers )
each of the terms in the sum is a number between zero and one. A reasonable
way to implement the neighbor list is to ignore terms less than
a small number (TOL) at every step.  On neighbor list update steps we recalculate
all the contributions to the sums.  Anything less than NL_TOL (a number less
than TOL) can safely be ignored until the next neighbor list update step.
For complex variables (e.g. q6) a considerable speedup can be acchieved just by
setting the value of TOL.  However, for variables such as coordinationnumbers,
where calculating all the distances is the most expensive part of the calculation,
the neighbor list and a TOL value are required to get the code to run fast