Made multicolvar and MultiColvarFunction inherit from a common base class
This makes these two classes slightly simpler.  It also allows one to
use MultiColvarFunctions on MultiColvarFunctions.  We can thus do things
like the value of the dot product of the q4 vector on a atom with the
q4 vectors on the atoms in the first coordination sphere averaged over a
subregion of the simulation box.