Changed routine for adding forces and made it so that you can use numerical derivatives everywhere
To get the forces on the vessels you now call getForcesOnVessels this returns
a bool which tells you if you need to call ActionAtomistic::setAtomicForces or
ActionWithArguments::addForcesToArguments.  Separately numerical derivatives
were a problem if you had an action that inherits from both ActionAtomistic
and ActionWithArguments.  This is now not a problem as you can, in a manner of
speaking, specify where the atomic derivatives start when you call
ActionAtomistic::calculateNumericalDerivativea (i.e. after the derivatives wrt
to the Arguments )