Created a class to hold the vector of ReferenceConfigurations
A class is used to hold the vector of ReferenceConfiguration and the
points where each point is mapped and the weight.  This means that we
can reuse this functionality in analysis for creating MDS and these sort
of things.  In addition, you can do dimensionality reduction with analysis
and then use the result in mapping.