Modified use of neighbor list in COORDINATION
It solves a known problem with the calculation
of acceptance in replica exchange with COORDINATION
and NLIST.

Notice that it changes the behavior of neighbor list
in the first two steps of a simulation. Previously,
during those steps full pair list was used. Now,
first step is equivalent to any other step
multiple of NL_STRIDE. Because of this, I updated
a couple of regtests.