Replaced VolumeSubcell with VolumeAround
The old VolumeSubcell did not have translational invariance.  This
has been resolved in around by adding a reference atom about which
we integrate the density. VolumeSubcell has been removed as it is not
yet working.  Also changes were required in vesselbase
to make the numerical derivatives of these "bridged" quantities work.