Fixed bugs in documentation
For some reason on my local linux machine doxygen does not
replace the variables $extrastylesheet and $relpath in the
PlumedUserHeader.html file.  As we are not using these variables
I deleted them from the header files.  I think this will cause problems
if we ever choose to use a custom css file for the documentation or if
we use the CREATE_SUBDIRS option.  I cannot imagine that we will ever
use these options but I thought I would mention these problems in the
log.