Added framework to use driver on multiple trajectories
I adapted the debug replica exchange stuff. Since running on multiple
trajectory could be useful, I leave --multi as a non-hidden option.

I slightly changed behavior of the hidden flag --debug-grex (now one sets the exchange stride, not the number of replicas - the latter is set with --multi). Thus, a couple of retests had to be updated