Simplification of storage of data in multicolvar
For a number of complex multicolvars eg. Local_q4 it is useful
to store the quantities calculated in one multicolvar and to use them later.
I had written functionality for this in a somewhat ad hoc manner intially and
as such there was much duplication of code and non uniformity in implications.
This resolves these issues.  Data is now stored more consistently and it is possible
to use a low memory version of this functionality at all times.