Fixed small things in analysis
Made it so that virtual routine is not called from constructor.  If the calculation
is terminated during the analysis phase then the analysis is now restarted when
the calculation reaches update on the first loop.  I also removed the duplicated
backup code that was in analysis.  Backup is now done by the routines in OFile.
This routine is now a little more flexible in that you can specify the string
in the backup name.