Added regtest type=make
It allows writing regtests to verify consistency of the plumed tool library.
It is designed in a way such that it can both use the "sourceme.sh" version
of plumed or the installed one, by checking if plumed exec is installed
or not.
I also added an example regtest - which should likely be expanded.