Added --save-originals to patch
I added a new mode to store patches. If one saves them with
--save-originals (instead of -s) what happens is that
/patches/codename.diff
will be a directory. Instead of storing diff files there,
we will store original/modified pairs.

When patching, if a directory is found in
/patches/codename.diff
the diff are computed and then reapplied on the actual files.

The only change is that the diff is not stored but re-computed on-the-fly.
This has a completely equivalent effect of the original way.
E.g., one can patch two slightly different versions
of an MD code with the same patch.

However, it has the advantage that stored files are more friendly.
E.g., one could directly edit them. Also the git-merge of two
different patches should work better now.