Improved RMSD with align!=displace
I did it in the boring way - full derivatives of rotation matrix.
It seems to work now, but still breaks some test. It is probably
some numerical issue, I will investigate before merging.

It is slower than the align==displace version (~ a factor 1.5),
but I think it is more than acceptable.