Fixed stopflag in Driver
I fixed the usage of stopflag in driver so that it is clearer
that the program stops at the end of the present step and not
at the beginning of the following. This is obviously equivalent,
but shows that scope of the flag can be reduced to be defined
at the inner of the MD loop.

Even if this is not a bux fix, I consider it as paired with fix #47
and I merge it to v2.0