Added possibility to use NLIST with replex
I solved it in the simple way. Instead of asking gromacs
if the exchange has been done or not, I just recompute the neighbor
list explicitly at the step after exchange. This is obtained
by pretending we are on first step in the step after exchange.

Addresses #31 for the case of COORDINATION