Small fix in plumed mklib
Target shared library is deleted before attempting its creation.
In this manner, if an error arises during compilation, the
library will not be there.

It is important to have LOAD FILE=xxx.cpp working properly.
E.g. if there is an error during compilation, a possibly existing
library is deleted.