Changed Dependencies from set to vector
This makes it such the order in which dependencies are activated
is deterministic. It is related to a bug (#51) which was already fixed
by changing the way prepare() works in multicolvar. Anyway, I commit
this change because I think it is clearer if, in a multiple proc simulation,
all the procs prepare all the actions in the very same order.