Fields with cubic interpolation
This makes the field cvs fast enough to actually be used for production
calculations.  In addition, I have added an
action that allows you to use the interpolation objects to take in a function from a
grid that is calculated by some function and to interpolate said function onto a finer
grained grid.