Added flags to switch off library search (on by default)
All libraries can be switched off with
./configure --disable-name

This allows testing or removing libraries that are
known to be buggy or problematic.

Notice that when lapack/blas are off, internal versions
are used (plumed does not work without them).

Also notice that when mpi search is on standard mpi compilers
are searched for before non-mpi ones.
Thus, to configure with mpi just type:

./configure

To configure without mpi use:

./configure --disable-mpi

I made this change since mpi is commonly needed and there is no
expected drawback from using mpi.