Fix #79
I changed the behavior of METAD INTERVAL in such a way that the bias
is correctly set to the value at the boundary when the CV is outside
of the interval. This changes the behavior of plumed when using
METAD INTERVAL and any sort of replica exchange (e.g. bias exchange).

Before the fix the bias was incorrectly set to zero outside the
interval.

* NOTICE THAT THIS BEHAVIOR IS DIFFERENT FROM THAT IMPLEMENTED
  IN PLUMED 1*