Made single class that allows you to pass vessels
There are now a number of places where data in vessels is passed
between actions.  For instance this is done in multicolvar functions
such as local_q4 and when calculating the average value of the CV in
a particular volume.  I had done this in an ad-hoc way so there were
duplicated bits of code in different places.  I have now changed the
code so that there is less duplication of code.  There is now a single
base class which (most) of these vessel using actions inherit from.
MultiColvar functions still copy data across manually