Fixed regression introduced in 4894056f860
There was a problem when using plumed as an external library
(as in "g++ pippo.cpp -lplumed) in recursively included .h files.
The bug was appearing only when using installed version of plumed.
The reason was that recursively included files were not properly
installed.

I also add a note on this regression in the changelog of 2.0.2