Norm of average vector in volume implemented
I have added a vessel so that you can calculate the average vector
directly from a VectorMultiColvar.  This replaces the old AVERAGE_VECTOR
feature and is better because you can use this vessel in combination with
AROUND to calculate the norm of the average vector within a region of the box.