FuncSumHills and SumHills extension
added the options to set spacing instead of nbin and
the option to translate the feses with the minimum set
to zero.

This also fix #72.

More work is needed to make FuncSumHills fully compatible
with COLVAR files (i.e. adding the possibility of doing
histogram only on selected collective variables/fields instead
of using all the fields).