Fixes to the manual
Changes to the code are there so that I can reference the atoms
page in the descriptions of keywords that read atom numbers.  I also
changed the input to multicolvar functions.  I stopped using the ARG
keyword in these actions and used DATA instead because you are refering
to the data calculated in a multicolvar as opposed to a scalar quantity
calculated by a particular action.