Removed dynamic all_atoms from multicolvar
In multicolvar the list of atoms used to update when the
neighbor list changed.  I think this did very little for
the speed of the code.  What it certainly did was make
editing confusing.  I have thus removed this feature and
made it so that the list of atom indexes is hard set during
construction and is fixed for the duration of the run.