These modules should not be compulsory
plumed core can be compiled without these.

With this change it is necessary to solve a but in the script building the
list of modules to be tested. I took the occasion to simplify it.
Now, all the optional modules are checked, but if a directory is
not present in regtest/ that module is ignored.