PDF manual
Added PDF manual. I had to reorganize a bit the documentation to make
it more linear. In particular, I assigned each of the previously
"dangling" pages as subpage of another one (e.g. DEBUG keyword).
Additionally, I changed a bit the way MD-specific instructions are linked
and added some comment about the PDF manual in the introduction.

The most important change is that I am not using the Layout anymore since
it was difficult to keep it in sync. Now, in the html version, only
the left-side tree is visible, and the tabs disappeared. I don't think
this is a big issue but we might try to solve it.

PDF manual is still a bit buggy: some links are not properly resolved
and images are not shown. I think we can solve these issues later.

Addresses #101