Implemented UPDATE_FROM and UPDATE_UNTIL
When specified, the update() method is only invoked
within this time range. This allows to perform some tasks
in a single simulation, e.g.:
Equilibration + Metadynamics + umbrella sampling refinement
with analysis of the last part only.

Currently this is an optional flag for all the actions.
I am not sure this is a good idea (currently update() method
is only used in some ActionPilot), since this will make
the keyword appear in every manual page.
However, since the update method is in principle available on all
the actions I preferred to add these keywords everywhere.

Fixes #65