Properly append suffix to xtc/trr files.
Since now plumed can write xtc/trr files, in a multi replica
framework I think it makes more sense to add the replica suffix before
the extension. That is, with DUMPATOMS FILE=test.xtc, we get
files such as test.0.xtc and not test.xtc.0

The same happens for trajectories read from the driver.

Notice that this was already true for .gz files, and now applies
for .xtc and .trr files. All other extensions are still ignored
for backward compatibility.