Changed analysis so that data are accumulated in update()
This allows me to have UPDATE_FROM and UPDATE_TO working properly
with histogram.

@gtribello you might want to check if this is ok.
Notice that without this fix histograms accumulated in a replica exchange
simulation would include also data from other replicas, which is a bug.
I thus think this commit should be applied also to v2.0 and v2.1.