Fixed portability of Stopwatch
I fixed some probability issue in Stopwatch.
Before this fix, the macro used to check the presence of
clock_gettime was not properly recognized. Additionally,
gettimeofday was considered as compulsory, causing problems
on bgxlc (as pointed out by Teodoro Laino).

After this fix, both gettimeofday and clock_gettime
are searched and considered as optional. If present,
clock_gettime is used. Otherwise gettimeofday is used.
If none is present, time is just not measured.